Added basic admin login system + new pages.
-Added basic admin login system.

-Now only admins can access or make        any changes to the database as there are multiple safeguards within the code that check for a logged in admin.

-Added a login page (This can be used for user logins if added at a later date)

-Added an Admin Control Panel page.

-Removed the Database navigation options from the Navbar.

+++ src/app/db-get/db-get.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import Item from '../Item';
 import { ItemService } from '../item.service';
+import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
   selector: 'app-db-get',
@@ -10,24 +11,44 @@ import { ItemService } from '../item.service';
 export class DbGetComponent implements OnInit {
 
   items: Item[];
+  adminLoggedIn = window.localStorage.getItem('adminLoggedIn');
 
-  constructor(private is: ItemService) { }
+  constructor(private is: ItemService, private route: ActivatedRoute, private router: Router) { }
 
   ngOnInit() {
+    // First check if the admin is logged in.
+    if (this.adminLoggedIn === 'false') {
+      this.router.navigate(['home']);
+      console.log('NOT ADMIN');
+    } else {
+      console.log('IS ADMIN');
+    }
     this.getItems();
     }
 
   getItems(): void {
-    this.is.getItems().subscribe((data: Item[]) => {
-    this.items = data;
-    });
+    // We only want admins to make changes.
+    if (this.adminLoggedIn === 'true') {
+      this.is.getItems().subscribe((data: Item[]) => {
+        this.items = data;
+        });
+    } else {
+      window.alert('You are not an admin!');
+      this.router.navigate(['home']);
+    }
   }
 
   deleteItem(id) {
+    // We only want admins to make changes.
+    if (this.adminLoggedIn === 'true') {
     this.is.deleteItem(id).subscribe(res => {
       console.log('Deleted item: ' + id);
       this.getItems();
       });
+    } else {
+      window.alert('You are not an admin!');
+      this.router.navigate(['home']);
+    }
   }
 
 }
